[CHANGE] `models` directory structure
This is a breaking change.
Solve an error in `SQLModel` dependency cased by recursive importations
while working on models.
