+++ app/models/__init__.py
@@ -1,6 +1,6 @@
 __all__ = ["Permission", "Role", "RoleHasPermission", "User"]
 from app.models.auditor_model import Auditor
-from app.models.permission_model import Permission
-from app.models.role_has_permission_model import RoleHasPermission
-from app.models.role_model import Role
-from app.models.user_model import User
+from app.models.models import Permission
+from app.models.models import RoleHasPermission
+from app.models.models import Role
+from app.models.models import User
